+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRMVC5WPZR25A5MJQFYZ8H71"
+        "01KRNNJNFVB6T85QDTA31ZHBBY"
     )
     r"""The unique identifier of the data source"""